+++ lldb/source/Plugins/Process/Utility/RegisterInfoPOSIX_arm64.h
@@ -34,6 +34,7 @@ public:
     eRegsetMaskZT = 64,
     eRegsetMaskFPMR = 128,
     eRegsetMaskGCS = 256,
+    eRegsetMaskPOE = 512,
     eRegsetMaskDynamic = ~1,
   };
 
@@ -118,6 +119,8 @@ public:
 
   void AddRegSetGCS();
 
+  void AddRegSetPOE();
+
   uint32_t ConfigureVectorLengthSVE(uint32_t sve_vq);
 
   void ConfigureVectorLengthZA(uint32_t za_vq);
@@ -138,6 +141,7 @@ public:
   bool IsTLSPresent() const { return m_opt_regsets.AnySet(eRegsetMaskTLS); }
   bool IsFPMRPresent() const { return m_opt_regsets.AnySet(eRegsetMaskFPMR); }
   bool IsGCSPresent() const { return m_opt_regsets.AnySet(eRegsetMaskGCS); }
+  bool IsPOEPresent() const { return m_opt_regsets.AnySet(eRegsetMaskPOE); }
 
   bool IsSVEReg(unsigned reg) const;
   bool IsSVEZReg(unsigned reg) const;
@@ -151,6 +155,7 @@ public:
   bool IsSMERegZT(unsigned reg) const;
   bool IsFPMRReg(unsigned reg) const;
   bool IsGCSReg(unsigned reg) const;
+  bool IsPOEReg(unsigned reg) const;
 
   uint32_t GetRegNumSVEZ0() const;
   uint32_t GetRegNumSVEFFR() const;
@@ -164,6 +169,7 @@ public:
   uint32_t GetSMEOffset() const;
   uint32_t GetFPMROffset() const;
   uint32_t GetGCSOffset() const;
+  uint32_t GetPOEOffset() const;
 
 private:
   typedef std::map<uint32_t, std::vector<lldb_private::RegisterInfo>>
@@ -197,6 +203,7 @@ private:
   std::vector<uint32_t> m_sme_regnum_collection;
   std::vector<uint32_t> m_fpmr_regnum_collection;
   std::vector<uint32_t> m_gcs_regnum_collection;
+  std::vector<uint32_t> m_poe_regnum_collection;
 };
 
 #endif